+++ src/util/tradeEventManagement.ts
@@ -21,7 +21,7 @@ import {
 import { logger } from "./logger";
 import { CleanStartEvent } from "../types/typechain/IMangrove";
 
-type Optional<T, K extends keyof T> = Pick<Partial<T>, K> & Omit<T, K>;
+export type Optional<T, K extends keyof T> = Pick<Partial<T>, K> & Omit<T, K>;
 
 export type OrderResultWithOptionalSummary = Optional<
   Market.DirtyOrderResult,